add module nesting information in the retriever
